quick error printing
